Abstract out the IPC client, so other clients can be used
