+++ ipc/src/clients/ipc/mod.rs
@@ -1,0 +1,2 @@
+pub mod ipc_client;
+pub mod ipc_socket;
