+++ launcher/src/main.rs
@@ -5,9 +5,9 @@ use std::ffi::CString;
 use anyhow::{bail, Result};
 use std::path::PathBuf;
 
-
 use goxlr_ipc::client::Client;
-use goxlr_ipc::ipc_socket::Socket;
+use goxlr_ipc::clients::ipc::ipc_client::IPCClient;
+use goxlr_ipc::clients::ipc::ipc_socket::Socket;
 use goxlr_ipc::{DaemonRequest, DaemonResponse};
 use interprocess::local_socket::tokio::LocalSocketStream;
 use interprocess::local_socket::NameTypeSupport;
@@ -98,7 +98,7 @@ async fn open_ui() -> Result<()> {
 
     if let Some(connection) = usable_connection {
         let socket: Socket<DaemonResponse, DaemonRequest> = Socket::new(connection);
-        let mut client = Client::new(socket);
+        let mut client = IPCClient::new(socket);
         client.send(DaemonRequest::OpenUi).await?;
         return Ok(());
     }
